refactored edit screen to use "CaptionStyleObject" instead of individual caption style properties
