+++ src/types/video.js
@@ -6,7 +6,7 @@ export type CaptionTextAlignment = 'center' | 'left' | 'right';
 export type CaptionWordStyle = 'animated' | 'none';
 export type CaptionBackgroundStyle = 'gradient' | 'solid';
 
-export type CaptionPresetStyleObject = {
+export type CaptionPresetStyleObject = {|
   textAlignment: CaptionTextAlignment,
   lineStyle: CaptionLineStyle,
   wordStyle: CaptionWordStyle,
@@ -14,9 +14,9 @@ export type CaptionPresetStyleObject = {
   backgroundColor: ColorRGBA,
   fontFamily: string,
   textColor: ColorRGBA,
-};
+|};
 
-export type CaptionStyleObject = {
+export type CaptionStyleObject = {|
   textAlignment: CaptionTextAlignment,
   lineStyle: CaptionLineStyle,
   wordStyle: CaptionWordStyle,
@@ -25,4 +25,4 @@ export type CaptionStyleObject = {
   textColor: ColorRGBA,
   fontSize: number,
   fontFamily: string,
-};
+|};
